backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
anthropic>=0.39.0
tavily-python>=0.5.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
